Add check magic command
