--- jupyter-config/notebook.d/jupyterlab_requirements.json
@@ -1,5 +1,0 @@
-{
-    "load_extensions": {
-      "jupyterlab_requirements": true
-    }
-}
